Commit.getParents working with merge commits
getParents failed when more than one commit was present, since it attempted
to get the second parent of commit's parent instead of it's own second
parent.

The promise for the second parent was also pushed onto the array for
Promise.all() too late to actually have an effect.

Limit was not working properly as a limit, since it attempted to
retrieve the set limit regardless of how many parents a commit had,
failing if the limit was set to more than Commit.parentcount().