Started getting callback functions working in classes
Unlike structs, classes don't store the callbacks on the object.
Instead they are passed into the function and have to be hanled
differently. Currently this is breaking the combyne templating
and after that is fixed this should build. But this is what we
need to get going for status and diff to start working.