Skip transfer progress test until it's fixed
Currently the transfer progress test is breaking on windows
randomly. @tbranyen and I have been looking into it and it seems
like a libuv issue. I'm going to skip this now so that other
potential PR's aren't held up by it until we can get a good
resolution and fix it in another PR