Convert Clone from object to function
I find it really annoying that we have to do `NodeGit.Clone.clone` just
to clone a repository.  It would be much nicer to get a similar effect
like in C, where you can run `git_clone`.  This copies over the effect
of an object, but actually using our `clone` patched function instead.

I've made this commit first to show it does not introduce regressions
with our current clone code.