Fix `getHeadCommit` dying on empty repo
If you open a repo in NodeGit that's empty (i.e. has no commits yet)
doing `Repository.prototype.getHeadCommit` will error in the promise
chain. I think that really should just return null instead of killing
the promise chain.

There are also probably more of these in the library and we'll have
to add them as we find them.