Add `RevWalk.prototype.fastWalk`
This adds the manually written C++ function `fastWalk` the `RevWalk` class.

I also added in a system the let the developer add in optimized C++ code that will be added to
the library in a similar way that we wrap the libgit2 library. Basically, the developer just defines the 
API for the function to wrap and then just writes the C++ implementation of the functions. The
generator classes take care of the rest.