Make new template for shared result/error async return values
When git_graph_descendant_of returns, it returns a mixed result / error code. If the result is less than 0, it is an error, if it is 0 or higher, it is a result. We can now support this as part of the template so error handling is properly associated with this type of method. Set "isResultOrError": true in return in descriptor now.