Modify `ConvenientLine` to work with unicode
This changes how `ConvenientLines` are calculated with a breaking API change.

`ConvenientLine` used to assume that all content coming back was ASCII characters and
a uniform byte length. This is no longer the case. Now we use the byte length coming
back from libgit2 to extract the string from a buffer using the correct byte length. This
will give back the correct line everytime.

The breaking API change is that we no longer return the new line character with the line
nor do we include that character in the `contentLen` calculation either. I'm on the fence
about that change but it seems like it's what we ought to be doing instead of force all of
the consumers of the library to account for that everywhere.