First attempt at a fix for unallocated raw values
This will only handle string defaults, to satisfy the test, although
moving forward we should think about reasonable defaults for all
uninitialized properties.

Another potential issue with this is that I'm treating all strings as
uninitialized, which is probably not what we want, or is it? I could
envision a supplemental property that indicates this value should be
NULL initialized on `New` of its respective type.