Fix "commit on head on empty repo with createCommitOnHead" test
So this test was never valid ¯\_(ツ)_/¯

You can't call `Index#addByPath` with an absolute path which was what was happening. Since the method was sync the result was considered the actual result of the call and not the error code so this error was never caught. Now that the function is async, that result is correctly interpretted and the test failed correctly.

Changing this test to use the relative path fixes the test.