Fix submodule check
When installing NodeGit from npm as a dependency if the user had to build the binary it would fail on the submodule check. Now we just bypass that check on and error since the package would be published with the correct version of libgit2 in it and we don't need to grab it as a submodule at that point anyways so the check and logic is useless.