Normalize fn names for overrides
When we override a function on a nodegit type, what we generally do is pull the function out and save it, and write a new function to the object that calls the old saved value.
The names we used to save things were never consistant. Using the made up reset.getCommit as an example, we might have `originalGetCommit`, `_getCommit`, `resetGetCommit`, or if it wasn't a keyword, just `getCommit`. I changed them all to `_getCommit`. Also sometimes they would be at tht eotp of the file, sometimes down by the override. I moved them all up top.