Make `Branch.createFromAnnotated` async
The function was originally synchronous which meant it would return
something on regardless of whether it succeeded (a branch) or failed
(an error code). This forces the client to have to check whether the
returned object is an error code or an object. Change the function to
be asynchronous instead so that clients can use the resolved or
rejected promise to easily determine whether the branch was
successfully created or not.