Provide target repository to write the index to
The returned Index from Merge.commits() doesn't know where to write
the index to because it is not backed by a repository. This prevents
the example from running because the write() fails as it's only
in-memory and has no notion of a target .git/index file. Skip the
write() and just use writeTreeTo() immediately instead to write the
index to the sample repository.