Initialize pointers to prevent warnings during initial assignment
In convert_from_v8.cc, the unwrapping of values uses two separate lines
of code for variable declaration and variable assignment. When a
pointer (or double pointer) is present, the assignment to the newly
declared variable creates a compiler warning. To prevent the
compilation logs from filling up unnecessarily, pointer variables
should initially be assigned a `NULL` value upon its declaration so
that the compiler will not emit a warning on the subseqent variable
assignment line.