Add JavaScript version of `git_tag_peel()`
The git_tag_peel() function from libgit2 naturally takes a pointer as
a parameter for providing the caller with the peeled object. However,
this kind of API does not make sense in the JavaScript world so we
need to change the generated code to return the actual object to the
caller.