Change function calls to field accesses instead
In details-for-tree-entry.js, there are several calls to functions of
an IndexEntry. However, those are actually fields on an object
instead of prototype functions so referring to them as functions will
cause a ReferenceError to be thrown. The fix is to not refer to them
as functions but to simply refer to them as fields instead.

Signed-off-by: Remy Suen <remy.suen@gmail.com>