Change tree_entry to not be selfFreeing
The libgit2 use cases all are instances where the git_tree_entry is owned by the object that returned it. As such this marks the tree_entry as not self freeing. And also updates functions returning one of these to set ownedByThis. This will keep the owning object alive should someone have a reference to a tree_entry.

The git_treebuilder_insert method needed to be marked as not async in order to have access to the owning object. This seems like a case where the generator could be improved should this method need to be async.