Expose JavaScript function name in thrown error
When an error happens in libgit2, NodeGit will wrap the and expose
the error code and message to be consumed via the JavaScript API.
However, this information is often insufficient given the
asynchronous nature of NodeGit and it may at times be unclear as to
which JavaScript function caused the error on the libgit2 side. By
exposing an 'errorFunction' property string that includes the name of
the JavaScript function in the error, debugging should now be a lot
easier as clients will now be able to quickly identify which
JavaScript function in a long promise chain was the source of an
error.

Signed-off-by: Remy Suen <remy.suen@gmail.com>