Generate options if unspecified in Revert.revert
Due to the way the C++ binding layer gets the promise callback passed
to it, if there are no options specified in Revert.revert, the
promise callback ends up becoming the third parameter to the function
and the expected fourth parameter will not be satisfied which will
lead to an exception. We should generate a default options for
Revert.revert if nothing is provided by the callee to make the API
easier and simpler to use.

Signed-off-by: Remy Suen <remy.suen@gmail.com>