Output warning if unmatched arg name used in generator
When making additions to the NodeGit API, it is necessary to make
changes to the descriptor.json. Small typos can cause the generated
code to not match the expected output. We should generate a warning
if an argument name used in descriptor.json does not match the name
in the C function. This will help make it a lot easier to discover
problems in the handwritten descriptor.json file.

Signed-off-by: Remy Suen <remy.suen@gmail.com>