Encode test keys using base64 so that github doesn't flag them as bad
We've encoding the 2 test private keys that we use in base 64 and we decode them before running tests. GitHub has begun auditing repos for ssh keys and flagging them as unsafe for use if they are in public repos. This is a very good thing on part of GitHub, but bad for our test apparatus.