Workaround: Prevent objectwrap from being cleaned up during async work
We're seeing some super weird behavior where an object that was persisted is being cleaned up during async work. We need to workaround this issue, as we are reasonably certain that our persistent references are well formed. We suspect this is an issue with v8 itself rather than an issue with our code. We are now aggressively preventing GC from calling the objectwrap destructor until the objectwrap is no longer in use.