Free memory when closing context
There was a memory leak when v8 context is closed: the Nan::ObjectWrap objects created inside nodegit were not freed, since their WeakCallback were not triggered.

To fix it, a new class has been added from which NodeGitWrapper objects inherit, so that they can be tracked in the nodegit::Context and freed when its cleanup handle is triggered.

This new class allows for cheap tracking via Link/Unlink to a tracker list, allowing each NodeGitWrapper object to link/unlink itself while the normal GC is running, as well as for the nodegit::Context to keep a tracker list of them and free them in a 'owned objects first' way when closing.

The ownership model and the fact that a NodeGitWrapper object might have multiple owners have also been considered.