Unsafe temporary workaround for LFS checkout lost performance
This is a temporary workaround in order to avoid the lost of performance with LFS checkout.

The change is limited to the processing of callbacks from Workers that leverage threaded libgit2 functions. Basically what it does is allowing the callbacks from executorEventsQueue to be queued in jsThreadCallbackQueue without waiting for the current one to end.

It is unsafe because with threaded libgit2 functions there is a potential risk of deadlock if the callbacks need to lock an object.

This commit will be reverted when nodegit-lfs is integrated into nodegit.