don't duplicate repo when getting remotes
I didn't realize this beforehand but the git_remote object takes the pointer to the repo it's looked up on. Then when we construct a GitRemote, we create a wrapper around *that* repo pointer(because we create a new repository instance per-remote...) which gets garbage collected later. I still don't see a good reason to duplicate the repo pointer and we don't in other cases of constructing GitRemote so I'm gonna remove it.